refactor(agent-task): split runner-exec tests out of submit_and_persist (#9927)
