+++ crates/homeboy-agents/src/agent_task_lifecycle/tests/mod.rs
@@ -382,6 +382,7 @@ pub(super) fn succeeded_aggregate(plan: &AgentTaskPlan) -> AgentTaskAggregate {
 mod handoff_and_proxy;
 mod operation_claims;
 mod private_attachment;
+mod runner_exec;
 mod status_and_recovery;
 mod submit_and_persist;
 mod terminal_and_reconcile;
